refactor(ui): move chat hooks to headless
